Fixes for schedule and scheduleDetail management.

+++ LMSProject/Controllers/CalendarController.cs
@@ -10,6 +10,12 @@ using System.Data.Entity;
 
 namespace LMSProject.Controllers
 {
+    public class ManageScheduleDetailsViewModel
+    {
+        public IEnumerable<scheduleDetail> details;
+        public int scheduleID;
+    }
+
     public class CalendarViewModel
     {
         public int Hours { get; set; }
@@ -110,6 +116,7 @@ namespace LMSProject.Controllers
         [HttpPost]
         public ActionResult CreateScheduleDetail(scheduleDetail detail)
         {
+            int detailID = detail.scheduleDetailID;
             ViewData["taskID"] = new List<SelectListItem> {
                 new SelectListItem { Text = "9A", Value = "1"},
                 new SelectListItem { Text = "9B", Value = "1"}
@@ -121,7 +128,7 @@ namespace LMSProject.Controllers
 
             if (ModelState.IsValid)
             {
-                if (IsOverlappingWithAnother(detail))
+                if (IsOverlappingWithAnother(detail,detailID))
                 {
                     return RedirectToAction("ErrorOverlapping");
                 }
@@ -133,7 +140,7 @@ namespace LMSProject.Controllers
         }
 
         // GET: schedules/Create
-        public ActionResult Create()
+        public ActionResult CreateSchedule()
         {
             ViewBag.schoolClassID = new SelectList(db.schoolClasses, "schoolClassID", "name");
             return View();
@@ -144,7 +151,7 @@ namespace LMSProject.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "scheduleID,schoolClassID")] schedule schedule)
+        public ActionResult CreateSchedule([Bind(Include = "scheduleID,schoolClassID")] schedule schedule)
         {
             if (ModelState.IsValid)
             {
@@ -160,15 +167,15 @@ namespace LMSProject.Controllers
         [HttpGet]
         public ActionResult CreateScheduleDetail(int id)
         {
-            scheduleDetail detail = new scheduleDetail() { startTime = DateTime.Today.AddHours(9), name = "Test", endTime = DateTime.Today.AddHours(12) };
+            scheduleDetail detail = new scheduleDetail() { startTime = DateTime.Today.AddHours(9), name = "", endTime = DateTime.Today.AddHours(12), scheduleID = id};
             ViewData["taskID"] = new SelectList(db.tasks.Where(p => p.schoolClassID == id).ToList(), "taskID", "name");
-            ViewData["scheduleID"] = new SelectList(db.schedules, "scheduleID", "schoolClasses.name");
+            ViewData["scheduleID"] = new SelectList(db.schedules, "scheduleID", "schoolClasses.name", detail.scheduleID);
 
             return View(detail);
         }
 
         // GET: Calendar/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult DeleteScheduleDetail(int? id)
         {
             if (id == null)
             {
@@ -183,9 +190,9 @@ namespace LMSProject.Controllers
         }
 
         // POST: Calendar/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost, ActionName("DeleteScheduleDetail")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteScheduleDetailConfirmed(int id)
         {
             scheduleDetail scheduleDetail = db.scheduleDetails.Find(id);
             db.scheduleDetails.Remove(scheduleDetail);
@@ -214,13 +221,40 @@ namespace LMSProject.Controllers
         public ActionResult DeleteScheduleConfirmed(int id)
         {
             schedule schedule = db.schedules.Find(id);
+            if (schedule == null)
+            {
+                return HttpNotFound();
+            }
+            List<scheduleDetail> details = db.scheduleDetails.Where(p => p.scheduleID == id).ToList();
+            foreach( var d in details )
+            {
+                db.scheduleDetails.Remove(d);
+            }
+            db.SaveChanges();
+
             db.schedules.Remove(schedule);
+
             db.SaveChanges();
             return RedirectToAction("Index");
         }
+        private bool IsOverlappingWithAnother(scheduleDetail a, int detailID)
+        {
+            var details = from d in db.scheduleDetails
+                          where d.scheduleID == detailID
+                          select d;
+            foreach (scheduleDetail b in details.ToList())
+            {
+                if ((a.startTime < b.endTime && b.startTime < a.endTime)
+                    && a.scheduleDetailID != b.scheduleDetailID)
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
 
         // GET: scheduleDetails/Edit/5
-        public ActionResult Edit(int? id)
+        public ActionResult EditScheduleDetail(int? id)
         {
             if (id == null)
             {
@@ -231,40 +265,31 @@ namespace LMSProject.Controllers
             {
                 return HttpNotFound();
             }
-            //ViewData["scheduleDetailID"] = new SelectList(db.scheduleDetails.Where(p => p.scheduleDetailID == id).ToList(), "scheduleDetailID", "schoolClasses.name", scheduleDetail.scheduleDetailID);
-            ViewData["scheduleID"] = new SelectList(db.schedules, "scheduleID", "schoolClasses.name");
+            ViewData["scheduleDetailID"] = new SelectList(db.scheduleDetails.Where(p => p.scheduleDetailID == id).ToList(), "scheduleDetailID", "schoolClasses.name", scheduleDetail.scheduleDetailID);
+            ViewData["scheduleID"] = new SelectList(db.schedules, "scheduleID", "schoolClasses.name", scheduleDetail.scheduleID);
             ViewData["taskID"] = new SelectList(db.tasks.Where(p => p.schoolClassID == scheduleDetail.schedules.schoolClassID).ToList(), "taskID", "name");
             ViewData["schoolClassID"] = new SelectList(db.schoolClasses, "schoolClassID", "name");
+
             return View(scheduleDetail);
         }
 
-        private bool IsOverlappingWithAnother(scheduleDetail a)
-        {
-            var details = from d in db.scheduleDetails
-                          where d.scheduleID == a.scheduleID
-                          select d;
-            foreach(scheduleDetail b in details.ToList())
-            {
-                if (b.endTime > a.startTime && b.startTime < a.endTime)
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
         // POST: scheduleDetails/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //public ActionResult EditScheduleDetail([Bind(Include = "scheduleID,starTime,endTime,name,room,taskID")] scheduleDetail scheduleDetail)
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "scheduleID,starTime,endTime,name,room")] scheduleDetail scheduleDetail)
+        public ActionResult EditScheduleDetail(scheduleDetail scheduleDetail)
         {
+            //int scheduleDetailID = scheduleDetail.scheduleDetailID;
+            //if (IsOverlappingWithAnother(scheduleDetail, scheduleDetailID))
+            //{
+            //    return Content("Error: That scheduleDetail overlaps with another in the schedule.");
+            //}
+
             if (ModelState.IsValid)
             {
-                if(IsOverlappingWithAnother(scheduleDetail))
-                {
-                    return Content("ErrorOverlapping");
-                }
                 db.Entry(scheduleDetail).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -273,10 +298,49 @@ namespace LMSProject.Controllers
             return View(scheduleDetail);
         }
 
+        // GET: schedules/EditSchedule/5
+        public ActionResult EditSchedule(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            schedule schedule = db.schedules.Find(id);
+            if (schedule == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.schoolClassID = new SelectList(db.schoolClasses, "schoolClassID", "name", schedule.schoolClassID);
+            return View(schedule);
+        }
+
+        // POST: schedules/EditSchedule/5
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditSchedule([Bind(Include = "scheduleID,schoolClassID")] schedule schedule)
+        {
+            if (ModelState.IsValid)
+            {
+                db.Entry(schedule).State = EntityState.Modified;
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            ViewBag.schoolClassID = new SelectList(db.schoolClasses, "schoolClassID", "name", schedule.schoolClassID);
+            return View(schedule);
+        }
+
+
         [HttpGet]
         public ActionResult ManageScheduleDetails(int id)
         {
-            return View(db.scheduleDetails.Where(x => x.scheduleID == id).ToList());
+            ManageScheduleDetailsViewModel viewModel = new ManageScheduleDetailsViewModel
+            {
+                details = db.scheduleDetails.Where(x => x.scheduleID == id).ToList(),
+                scheduleID = id
+            };
+            return View(viewModel);
         }
 
         public ActionResult ErrorOverlapping()
